Add Hateoas for User devices
- Extract user devices stuff from UserController into a UserDevicesController
